requirements.txt
asgiref==3.8.1
Django==5.1.2
<<<<<<< HEAD
=======
gunicorn==20.1.0
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
>>>>>>> origin/main
sqlparse==0.5.1
